requirements.txt
boto3
Flask
python-dotenv
opencv-python
numpy
